requirements.txt
fastapi[standard]
sqlalchemy
jinja2

itsdangerous
easyocr
pyjwt

